financial-document-analyzer-main/requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
fastapi==0.110.0
python-dotenv==1.0.0
uvicorn==0.29.0
langchain_community==0.0.31
unstructured==0.13.2
pypdf==4.1.0
python-multipart==0.0.9